% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_skeleton.R
\name{create_skeleton}
\alias{create_skeleton}
\title{create_skeleton}
\usage{
create_skeleton(df)
}
\arguments{
\item{df}{a dataframe of raw data}
}
\value{
a dataframe
}
\description{
Create a skeleton spreadsheet
}
