% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate}
\usage{
translate(df, input_format = NULL, mapper = NULL)
}
\arguments{
\item{df}{a dataframe of raw data}

\item{input_format}{the format of the raw data}

\item{mapper}{A dictionary dataframe (expected to be mapper from the vida library)}
}
\value{
a dataframe
}
\description{
Translate raw data from a site format to babel
}
